+++ core-bundle/src/Doctrine/Schema/DcaSchemaProvider.php
@@ -3,7 +3,7 @@
 /*
  * This file is part of Contao.
  *
- * Copyright (c) 2005-2016 Leo Feyer
+ * Copyright (c) 2005-2017 Leo Feyer
  *
  * @license LGPL-3.0+
  */
@@ -36,7 +36,9 @@ class DcaSchemaProvider
     }
 
     /**
-     * {@inheritdoc}
+     * Creates a schema.
+     *
+     * @return Schema
      */
     public function createSchema()
     {