+++ core-bundle/src/Resources/contao/widgets/SelectMenu.php
@@ -3,7 +3,7 @@
 /**
  * Contao Open Source CMS
  *
- * Copyright (c) 2005-2016 Leo Feyer
+ * Copyright (c) 2005-2017 Leo Feyer
  *
  * @license LGPL-3.0+
  */
@@ -118,7 +118,7 @@ class SelectMenu extends \Widget
 		}
 
 		// Add an empty option if there are none
-		if (empty($this->arrOptions))
+		if (empty($this->arrOptions) || !is_array($this->arrOptions))
 		{
 			$this->arrOptions = array(array('value'=>'', 'label'=>'-'));
 		}