+++ core-bundle/tests/Contao/StringUtilTest.php
@@ -19,6 +19,7 @@ use Contao\System;
  *
  * @author Yanick Witschi <https://github.com/toflar>
  * @author Martin Auswöger <martin@auswoeger.com>
+ * @author Leo Feyer <https://github.com/leofeyer>
  *
  * @group contao3
  */
@@ -44,6 +45,25 @@ class StringUtilTest extends TestCase
         }
 
         System::setContainer($this->mockContainerWithContaoScopes());
+    }
+
+    /**
+     * Tests the generateAlias() method.
+     */
+    public function testGenerateAlias()
+    {
+        $GLOBALS['TL_CONFIG']['characterSet'] = 'UTF-8';
+
+        $this->assertEquals('foo', StringUtil::generateAlias('foo'));
+        $this->assertEquals('foo', StringUtil::generateAlias('FOO'));
+        $this->assertEquals('foo-bar', StringUtil::generateAlias('foo bar'));
+        $this->assertEquals('foo-bar', StringUtil::generateAlias('%foo&bar~'));
+        $this->assertEquals('foo-bar', StringUtil::generateAlias('foo&amp;bar'));
+        $this->assertEquals('foo-bar', StringUtil::generateAlias('foo-{{link::12}}-bar'));
+        $this->assertEquals('foo-bar', StringUtil::generateAlias('föö-bär'));
+        $this->assertEquals('id-123', StringUtil::generateAlias('123'));
+        $this->assertEquals('123foo', StringUtil::generateAlias('123foo'));
+        $this->assertEquals('foo123', StringUtil::generateAlias('foo123'));
     }
 
     /**